refactor: apply 모듈 분리
