+++ src/main/java/org/ject/support/domain/apply/controller/ApplyController.java
@@ -1,12 +1,12 @@
-package org.ject.support.domain.recruit.controller;
+package org.ject.support.domain.apply.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.ject.support.common.security.AuthPrincipal;
+import org.ject.support.domain.apply.dto.ApplyTemporaryRequest;
+import org.ject.support.domain.apply.dto.ApplyTemporaryResponse;
+import org.ject.support.domain.apply.dto.SubmitApplicationRequest;
+import org.ject.support.domain.apply.service.ApplyUsecase;
 import org.ject.support.domain.member.JobFamily;
-import org.ject.support.domain.recruit.dto.ApplyTemporaryRequest;
-import org.ject.support.domain.recruit.dto.ApplyTemporaryResponse;
-import org.ject.support.domain.recruit.dto.SubmitApplicationRequest;
-import org.ject.support.domain.recruit.service.ApplyUsecase;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
