+++ src/main/java/org/ject/support/domain/apply/dto/ApplyTemporaryRequest.java
@@ -1,4 +1,4 @@
-package org.ject.support.domain.recruit.dto;
+package org.ject.support.domain.apply.dto;
 
 import java.util.List;
 import java.util.Map;
