+++ src/main/java/org/ject/support/domain/apply/service/ApplyUsecase.java
@@ -1,8 +1,8 @@
-package org.ject.support.domain.recruit.service;
+package org.ject.support.domain.apply.service;
 
+import org.ject.support.domain.apply.dto.ApplyPortfolioDto;
+import org.ject.support.domain.apply.dto.ApplyTemporaryResponse;
 import org.ject.support.domain.member.JobFamily;
-import org.ject.support.domain.recruit.dto.ApplyPortfolioDto;
-import org.ject.support.domain.recruit.dto.ApplyTemporaryResponse;
 
 import java.util.List;
 import java.util.Map;
