+++ src/main/java/org/ject/support/domain/tempapply/domain/TemporaryApplication.java
@@ -1,14 +1,11 @@
 package org.ject.support.domain.tempapply.domain;
 
 import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBTable;
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Map;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import org.ject.support.domain.apply.dto.ApplyPortfolioDto;
 import org.ject.support.domain.member.JobFamily;
-import org.ject.support.domain.recruit.dto.ApplyPortfolioDto;
 import org.ject.support.external.dynamodb.domain.CompositeKey;
 import org.ject.support.external.dynamodb.domain.EntityWithPrimaryKey;
 import org.ject.support.external.dynamodb.util.ApplyTemporaryPortfolioConverter;
@@ -17,6 +14,10 @@ import org.ject.support.external.dynamodb.util.MapConverter;
 import software.amazon.awssdk.enhanced.dynamodb.mapper.annotations.DynamoDbAttribute;
 import software.amazon.awssdk.enhanced.dynamodb.mapper.annotations.DynamoDbBean;
 import software.amazon.awssdk.enhanced.dynamodb.mapper.annotations.DynamoDbConvertedBy;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.Map;
 
 @DynamoDBTable(tableName = "temporary_application")
 @DynamoDbBean
